# Conflicts:
#	src/main/java/com/codegym/finwallet/controller/WalletController.java
#	src/main/java/com/codegym/finwallet/entity/Wallet.java
#	src/main/java/com/codegym/finwallet/service/WalletService.java
#	src/main/java/com/codegym/finwallet/service/impl/TransactionServiceImpl.java
#	src/main/java/com/codegym/finwallet/service/impl/WalletServiceImpl.java

+++ src/main/java/com/codegym/finwallet/controller/TransactionTypeController.java
@@ -19,7 +19,7 @@ public class TransactionTypeController {
 
     @PostMapping("/add-budget/{id}")
     public ResponseEntity<TransactionType> addBudgetToTransactionType(@PathVariable Long id,
-                                                                      @RequestParam float transactionBudget) {
+                                                                      @RequestParam double transactionBudget) {
         TransactionType updatedTransactionType = transactionTypeService.addBudgetToTransactionType(id, transactionBudget);
         return new ResponseEntity<>(updatedTransactionType, HttpStatus.OK);
     }
