+++ src/main/java/com/codegym/finwallet/service/TransactionTypeService.java
@@ -3,7 +3,7 @@ package com.codegym.finwallet.service;
 import com.codegym.finwallet.entity.TransactionType;
 
 public interface TransactionTypeService {
-    TransactionType addBudgetToTransactionType(Long id,float transactionBudget);
-    boolean isBudgetExceeded(TransactionType transactionType, float transactionAmount);
+    TransactionType addBudgetToTransactionType(Long id,double transactionBudget);
+    boolean isBudgetExceeded(TransactionType transactionType, double transactionAmount);
     void alertBudgetExceeded();
 }
