+++ src/main/java/com/codegym/finwallet/service/WalletService.java
@@ -14,5 +14,5 @@ public interface WalletService {
     Wallet findById(Long id);
     CommonResponse editWallet(WalletRequest walletRequest,Long id);
 //    CommonResponse transferMoney(TransferMoneyRequest transferMoneyRequest);
-    CommonResponse addMoneyToWallet(Long walletId, float amount);
+    CommonResponse addMoneyToWallet(Long walletId, double amount);
 }
